moved menu generation code to Sienar.Menus namespace
