+++ src/Sienar.Architecture.Web/Infrastructure/WebApplicationAdapter.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Sienar.Html;
+using Sienar.Menus;
 using Sienar.Plugins;
 
 namespace Sienar.Infrastructure;
