+++ src/Sienar.Plugin.Cms.Client/Ui/UserBadge.razor
@@ -1,4 +1,5 @@
 ﻿@inherits ServiceAwareClickHandlerBase
+@using Sienar.Menus
 @implements IDisposable
 @inject IMenuGenerator MenuGenerator
 @inject IUserAccessor UserAccessor
