+++ src/Sienar.Ui/Layouts/DrawerNavContent.razor
@@ -1,4 +1,5 @@
-﻿@inject ComponentProvider ComponentProvider
+﻿@using Sienar.Menus
+@inject ComponentProvider ComponentProvider
 
 <div class="flex-grow-1">
 	<OptionalComponent
