+++ src/Sienar.Ui/Ui/DashboardItem.razor
@@ -1,4 +1,6 @@
-﻿<div class="dashboard-item d-flex flex-column align-center flex-shrink-0 mt-4">
+﻿@using Sienar.Menus
+
+<div class="dashboard-item d-flex flex-column align-center flex-shrink-0 mt-4">
 	<MudIcon
 		Class="mb-4"
 		Icon="@Item.Icon"
