+++ src/Sienar.Ui/Ui/SidebarNavMenuLink.razor
@@ -1,4 +1,5 @@
-﻿@inherits ServiceAwareClickHandlerBase
+﻿@using Sienar.Menus
+@inherits ServiceAwareClickHandlerBase
 
 @if (Link.Sublinks is null)
 {
