+++ src/Sienar.Ui/Ui/TopNavMenuLink.razor
@@ -1,4 +1,6 @@
-﻿@if (!_hasSublinks)
+﻿@using Sienar.Menus
+
+@if (!_hasSublinks)
 {
 	<MudButton
 		Class="mr-4"
