+++ src/Sienar.Utils/Infrastructure/SienarAppBuilder.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Sienar.Configuration;
 using Sienar.Extensions;
 using Sienar.Html;
+using Sienar.Menus;
 using Sienar.Plugins;
 
 namespace Sienar.Infrastructure;
