+++ src/Sienar.Utils/Menus/DefaultMenuGenerator.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 using Sienar.Extensions;
 using Sienar.Security;
 
-namespace Sienar.Infrastructure;
+namespace Sienar.Menus;
 
 /// <exclude />
-public class MenuGenerator : IMenuGenerator
+public class DefaultMenuGenerator : IMenuGenerator
 {
 	private readonly IUserAccessor _userAccessor;
 	private readonly MenuProvider _provider;
 
-	public MenuGenerator(
+	public DefaultMenuGenerator(
 		IUserAccessor userAccessor,
 		MenuProvider menuProvider)
 	{
